06_Builder/tender_parser/requirements.txt
playwright==1.48.2
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
